pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dockpulse"
version = "0.1.0"
description = "Container Resource Profiler & Right-Sizer for Docker"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Hariharan Ragothaman" },
]
keywords = ["docker", "containers", "profiling", "right-sizing", "devops", "resource-optimization"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "docker>=7.0.0",
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "ruamel.yaml>=0.18.0",
    "jinja2>=3.1.0",
    "plotly>=5.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.4.0",
    "mypy>=1.8.0",
]

[project.scripts]
dockpulse = "dockpulse.cli:app"

[project.urls]
Homepage = "https://github.com/hariharanragothaman/dockpulse"
Repository = "https://github.com/hariharanragothaman/dockpulse"
Issues = "https://github.com/hariharanragothaman/dockpulse/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/dockpulse"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # ruff-specific rules
]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["dockpulse"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short --strict-markers"
